requirements.txt
about-time==4.2.1 ; python_version >= "3.10" and python_version < "3.11"
aiohttp==3.8.5 ; python_version >= "3.10" and python_version < "3.11"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "3.11"
alive-progress==3.1.4 ; python_version >= "3.10" and python_version < "3.11"
async-timeout==4.0.3 ; python_version >= "3.10" and python_version < "3.11"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "3.11"
certifi==2023.7.22 ; python_version >= "3.10" and python_version < "3.11"
charset-normalizer==3.2.0 ; python_version >= "3.10" and python_version < "3.11"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.11" and platform_system == "Windows"
datasets==2.14.4 ; python_version >= "3.10" and python_version < "3.11"
dill==0.3.7 ; python_version >= "3.10" and python_version < "3.11"
evaluate==0.4.0 ; python_version >= "3.10" and python_version < "3.11"
filelock==3.12.2 ; python_version >= "3.10" and python_version < "3.11"
frozenlist==1.4.0 ; python_version >= "3.10" and python_version < "3.11"
fsspec==2023.6.0 ; python_version >= "3.10" and python_version < "3.11"
fsspec[http]==2023.6.0 ; python_version >= "3.10" and python_version < "3.11"
grapheme==0.6.0 ; python_version >= "3.10" and python_version < "3.11"
huggingface-hub==0.16.4 ; python_version >= "3.10" and python_version < "3.11"
idna==3.4 ; python_version >= "3.10" and python_version < "3.11"
joblib==1.3.2 ; python_version >= "3.10" and python_version < "3.11"
multidict==6.0.4 ; python_version >= "3.10" and python_version < "3.11"
multiprocess==0.70.15 ; python_version >= "3.10" and python_version < "3.11"
numpy==1.25.2 ; python_version >= "3.10" and python_version < "3.11"
packaging==23.1 ; python_version >= "3.10" and python_version < "3.11"
pandas==1.5.3 ; python_version >= "3.10" and python_version < "3.11"
pyarrow==13.0.0 ; python_version >= "3.10" and python_version < "3.11"
pydantic==1.10.12 ; python_version >= "3.10" and python_version < "3.11"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "3.11"
pytz==2023.3 ; python_version >= "3.10" and python_version < "3.11"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "3.11"
regex==2023.8.8 ; python_version >= "3.10" and python_version < "3.11"
requests==2.31.0 ; python_version >= "3.10" and python_version < "3.11"
responses==0.18.0 ; python_version >= "3.10" and python_version < "3.11"
safetensors==0.3.3 ; python_version >= "3.10" and python_version < "3.11"
scikit-learn==1.3.0 ; python_version >= "3.10" and python_version < "3.11"
scipy==1.11.2 ; python_version >= "3.10" and python_version < "3.11"
six==1.16.0 ; python_version >= "3.10" and python_version < "3.11"
threadpoolctl==3.2.0 ; python_version >= "3.10" and python_version < "3.11"
tokenizers==0.13.3 ; python_version >= "3.10" and python_version < "3.11"
tqdm==4.66.1 ; python_version >= "3.10" and python_version < "3.11"
transformers==4.32.0 ; python_version >= "3.10" and python_version < "3.11"
typing-extensions==4.7.1 ; python_version >= "3.10" and python_version < "3.11"
urllib3==2.0.4 ; python_version >= "3.10" and python_version < "3.11"
xxhash==3.3.0 ; python_version >= "3.10" and python_version < "3.11"
yarl==1.9.2 ; python_version >= "3.10" and python_version < "3.11"
